Generated Control Rig First Pass
In Unreal 5.0 and later a Control Rig will be added next to the Skeletal Mesh for a character.  This is generated from a Python script and uses the new Full Body IK node.
This isn't a completed rig.
An anim file will no longer be created next to the Skeletal Mesh.  With Fixed Rotations as the default, this animation is built wrong anyway and leads to confusion.